awesome: made battery widget more robust
Some kernels do not provide a useful 'current_now' -- the file exists but the
device does not. This means that the widget thinks it can determine the rate
that the battery changes/discharges, but simply causes an error. Checking that
the rate is not 'nil' is insufficient in this case.